+++ packages/main/src/components/VariantManagement/index.tsx
@@ -1,7 +1,7 @@
 import '@ui5/webcomponents-fiori/dist/illustrations/UnableToLoad.js';
-import '@ui5/webcomponents-icons/dist/decline';
-import '@ui5/webcomponents-icons/dist/navigation-down-arrow';
-import '@ui5/webcomponents-icons/dist/search';
+import '@ui5/webcomponents-icons/dist/decline.js';
+import '@ui5/webcomponents-icons/dist/navigation-down-arrow.js';
+import '@ui5/webcomponents-icons/dist/search.js';
 import { useI18nBundle } from '@ui5/webcomponents-react-base/dist/hooks';
 import { StyleClassHelper } from '@ui5/webcomponents-react-base/dist/StyleClassHelper';
 import { ThemingParameters } from '@ui5/webcomponents-react-base/dist/ThemingParameters';