+++ packages/main/src/webComponents/AvatarGroup/index.tsx
@@ -4,7 +4,7 @@ import { CommonProps } from '@ui5/webcomponents-react/interfaces/CommonProps';
 import { Ui5CustomEvent } from '@ui5/webcomponents-react/interfaces/Ui5CustomEvent';
 import { ReactNode } from 'react';
 
-import '@ui5/webcomponents/dist/AvatarGroup';
+import '@ui5/webcomponents/dist/AvatarGroup.js';
 
 export interface AvatarGroupPropTypes extends Omit<CommonProps, 'onClick'> {
   /**