+++ packages/main/src/webComponents/CardHeader/index.tsx
@@ -3,7 +3,7 @@ import { CommonProps } from '@ui5/webcomponents-react/interfaces/CommonProps';
 import { Ui5CustomEvent } from '@ui5/webcomponents-react/interfaces/Ui5CustomEvent';
 import { ReactNode } from 'react';
 
-import '@ui5/webcomponents/dist/CardHeader';
+import '@ui5/webcomponents/dist/CardHeader.js';
 
 export interface CardHeaderPropTypes extends Omit<CommonProps, 'onClick'> {
   /**