+++ packages/main/src/webComponents/GroupHeaderListItem/index.tsx
@@ -2,7 +2,7 @@ import { withWebComponent } from '@ui5/webcomponents-react/dist/withWebComponent
 import { CommonProps } from '@ui5/webcomponents-react/interfaces/CommonProps';
 import { ReactNode } from 'react';
 
-import '@ui5/webcomponents/dist/GroupHeaderListItem';
+import '@ui5/webcomponents/dist/GroupHeaderListItem.js';
 
 export interface GroupHeaderListItemPropTypes extends CommonProps {
   /**