+++ packages/main/src/webComponents/TabContainer/index.tsx
@@ -4,7 +4,7 @@ import { CommonProps } from '@ui5/webcomponents-react/interfaces/CommonProps';
 import { Ui5CustomEvent } from '@ui5/webcomponents-react/interfaces/Ui5CustomEvent';
 import { ReactNode } from 'react';
 
-import '@ui5/webcomponents/dist/TabContainer';
+import '@ui5/webcomponents/dist/TabContainer.js';
 
 export interface TabContainerPropTypes extends CommonProps {
   /**