+++ packages/main/src/webComponents/TextArea/index.tsx
@@ -4,7 +4,7 @@ import { CommonProps } from '@ui5/webcomponents-react/interfaces/CommonProps';
 import { Ui5CustomEvent } from '@ui5/webcomponents-react/interfaces/Ui5CustomEvent';
 import { ReactNode } from 'react';
 
-import '@ui5/webcomponents/dist/TextArea';
+import '@ui5/webcomponents/dist/TextArea.js';
 
 export interface TextAreaPropTypes extends Omit<CommonProps, 'onChange' | 'onInput'> {
   /**