+++ CHANGELOG.md
@@ -3,6 +3,23 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.18.9](https://github.com/SAP/ui5-webcomponents-react/compare/v0.18.8...v0.18.9) (2021-09-21)
+
+
+### Bug Fixes
+
+* add `.js` file extension to all web components imports ([#2147](https://github.com/SAP/ui5-webcomponents-react/issues/2147)) ([2799eba](https://github.com/SAP/ui5-webcomponents-react/commit/2799eba322824f0995b48a31feadd4c43b889fed))
+* **MessageItem:** use correct display name ([#2129](https://github.com/SAP/ui5-webcomponents-react/issues/2129)) ([4357fab](https://github.com/SAP/ui5-webcomponents-react/commit/4357fabd9dbc169d07b15caa9438bbff436e2186))
+
+
+### Features
+
+* **MessageBox:** expose all `Dialog` props, except `footer`, `headerText` and `onAfterClose` ([#2141](https://github.com/SAP/ui5-webcomponents-react/issues/2141)) ([7a29323](https://github.com/SAP/ui5-webcomponents-react/commit/7a293231a8e5061e600f910cb30dbddd4c40751e))
+
+
+
+
+
 ## [0.18.8](https://github.com/SAP/ui5-webcomponents-react/compare/v0.18.7...v0.18.8) (2021-09-15)